Split WalkConfig class
