+++ plextraktsync/walker.py
@@ -17,72 +17,9 @@ if TYPE_CHECKING:
     from plexapi.video import Episode, Movie, Show
 
     from plextraktsync.media import Media, MediaFactory
+    from plextraktsync.plan.WalkConfig import WalkConfig
     from plextraktsync.plex.PlexApi import PlexApi
     from plextraktsync.plex.PlexLibrarySection import PlexLibrarySection
-
-
-class WalkConfig:
-    walk_movies = True
-    walk_shows = True
-    walk_watchlist = True
-    library = []
-    show = []
-    movie = []
-    id = []
-
-    def __init__(self, movies=True, shows=True, watchlist=True):
-        self.walk_movies = movies
-        self.walk_shows = shows
-        self.walk_watchlist = watchlist
-
-    def update(self, movies=None, shows=None, watchlist=None):
-        if movies is not None:
-            self.walk_movies = movies
-        if shows is not None:
-            self.walk_shows = shows
-        if watchlist is not None:
-            self.walk_watchlist = watchlist
-
-        return self
-
-    def add_library(self, library):
-        self.library.append(library)
-
-    def add_id(self, id):
-        self.id.append(id)
-
-    def add_show(self, show):
-        self.show.append(show)
-
-    def add_movie(self, movie):
-        self.movie.append(movie)
-
-    @property
-    def is_partial(self):
-        """
-        Returns true if partial library walk is performed.
-        Due to the way watchlist is filled, watchlists should only be updated on full walk.
-        """
-        # Single item provided
-        if self.library or self.movie or self.show or self.id:
-            return True
-
-        # Must sync both movies and shows to be full sync
-        return not self.walk_movies or not self.walk_shows
-
-    def is_valid(self):
-        # Single item provided
-        if self.library or self.movie or self.show or self.id:
-            return True
-
-        # Full sync of movies or shows
-        if self.walk_movies or self.walk_shows:
-            return True
-
-        if self.walk_watchlist:
-            return True
-
-        return False
 
 
 class WalkPlan(NamedTuple):
