+++ tests/test_walker.py
@@ -1,7 +1,8 @@
 #!/usr/bin/env python3 -m pytest
+from plextraktsync.plan.WalkConfig import WalkConfig
 from plextraktsync.plex.PlexApi import PlexApi
 from plextraktsync.plex.PlexLibrarySection import PlexLibrarySection
-from plextraktsync.walker import WalkConfig, WalkPlanner
+from plextraktsync.walker import WalkPlanner
 
 
 class PlexLibrarySectionMock(PlexLibrarySection):
